create db folder and include handlers and selectors folders
