+++ commands/bitpanda.ts
@@ -1,4 +1,4 @@
-import { handlers } from "../src/handlers/BITPANDA";
+import { handlers } from "../src/db/handlers/BITPANDA";
 import argv from "process.argv";
 import { PrismaClient } from "@prisma/client";
