+++ src/db/handlers/YOUNG_PLATFORM/buySellSwap.ts
@@ -1,27 +1,30 @@
 import { Prisma, PrismaClient } from "@prisma/client";
-import { convertCSVtoJSON } from "../../../convertCSVtoJSON";
+import { convertCSVtoJSON } from "../../../../convertCSVtoJSON";
 
 export type CsvInput = {
   id: number;
-  credit: number;
-  debit: number;
-  currency: string;
+  base: string;
+  quote: string;
+  amount: number;
+  volume: number;
+  rate: number;
+  brokerage: number;
+  brokerage_currency: string;
+  side: string;
   date: string;
-  tx_type: string;
 };
 
-export type Parsed = Omit<
-  Prisma.YoungPlatformMovementCreateInput,
-  "id" | "userAccountId"
->;
+type Parsed = Omit<
+  Prisma.YoungPlatformTradeCreateInput,
+  "userAccountId" | "id"
+> & { txnId: number };
 
 const parse = (input: CsvInput): Parsed => {
-  const { tx_type: txType, id, ...rest } = input;
-
+  const { brokerage_currency: brokerageCurrency, id, ...rest } = input;
   return {
     originalData: JSON.stringify(input),
-    moveId: id,
-    txType,
+    brokerageCurrency,
+    txnId: id,
     ...rest,
   } as unknown as Parsed;
 };
@@ -36,19 +39,19 @@ const store = async ({
   prisma: PrismaClient;
 }) =>
   Promise.all(
-    parsed.map(async ({ originalData, ...movement }) => {
-      console.log("adding deposit_withdraw_fee_order moveId", movement.moveId);
+    parsed.map(async ({ originalData, ...trans }) => {
+      console.log("adding buy_sell_swap txnId", trans.txnId);
       const data = {
-        ...movement,
+        ...trans,
         originalData: [originalData] as Prisma.JsonArray,
         userAccountId: userAccountId,
       };
-      await prisma.youngPlatformMovement.upsert({
-        where: { moveId: movement.moveId },
+      await prisma.youngPlatformTrade.upsert({
+        where: { txnId: trans.txnId },
         create: data,
         update: data,
       });
-      console.log("adding deposit_withdraw_fee_order moveId", movement.moveId);
+      console.log("added buy_sell_swap txnId", trans.txnId);
     })
   );
 
@@ -62,7 +65,7 @@ export const handle = async ({
   prisma: PrismaClient;
 }) => {
   const csvJsonData = await convertCSVtoJSON<CsvInput>(
-    `${year}/YOUNG_PLATFORM/deposit_withdraw_fee_order.csv`
+    `${year}/YOUNG_PLATFORM/buy_sell_swap.csv`
   );
   const parsed = csvJsonData.map(parse);
