+++ src/db/handlers/YOUNG_PLATFORM/depositWithdrawFeeOrder.ts
@@ -1,30 +1,27 @@
 import { Prisma, PrismaClient } from "@prisma/client";
-import { convertCSVtoJSON } from "../../../convertCSVtoJSON";
+import { convertCSVtoJSON } from "../../../../convertCSVtoJSON";
 
 export type CsvInput = {
   id: number;
-  base: string;
-  quote: string;
-  amount: number;
-  volume: number;
-  rate: number;
-  brokerage: number;
-  brokerage_currency: string;
-  side: string;
+  credit: number;
+  debit: number;
+  currency: string;
   date: string;
+  tx_type: string;
 };
 
-type Parsed = Omit<
-  Prisma.YoungPlatformTradeCreateInput,
-  "userAccountId" | "id"
-> & { txnId: number };
+export type Parsed = Omit<
+  Prisma.YoungPlatformMovementCreateInput,
+  "id" | "userAccountId"
+>;
 
 const parse = (input: CsvInput): Parsed => {
-  const { brokerage_currency: brokerageCurrency, id, ...rest } = input;
+  const { tx_type: txType, id, ...rest } = input;
+
   return {
     originalData: JSON.stringify(input),
-    brokerageCurrency,
-    txnId: id,
+    moveId: id,
+    txType,
     ...rest,
   } as unknown as Parsed;
 };
@@ -39,19 +36,19 @@ const store = async ({
   prisma: PrismaClient;
 }) =>
   Promise.all(
-    parsed.map(async ({ originalData, ...trans }) => {
-      console.log("adding buy_sell_swap txnId", trans.txnId);
+    parsed.map(async ({ originalData, ...movement }) => {
+      console.log("adding deposit_withdraw_fee_order moveId", movement.moveId);
       const data = {
-        ...trans,
+        ...movement,
         originalData: [originalData] as Prisma.JsonArray,
         userAccountId: userAccountId,
       };
-      await prisma.youngPlatformTrade.upsert({
-        where: { txnId: trans.txnId },
+      await prisma.youngPlatformMovement.upsert({
+        where: { moveId: movement.moveId },
         create: data,
         update: data,
       });
-      console.log("added buy_sell_swap txnId", trans.txnId);
+      console.log("adding deposit_withdraw_fee_order moveId", movement.moveId);
     })
   );
 
@@ -65,7 +62,7 @@ export const handle = async ({
   prisma: PrismaClient;
 }) => {
   const csvJsonData = await convertCSVtoJSON<CsvInput>(
-    `${year}/YOUNG_PLATFORM/buy_sell_swap.csv`
+    `${year}/YOUNG_PLATFORM/deposit_withdraw_fee_order.csv`
   );
   const parsed = csvJsonData.map(parse);
